pyproject.toml
[tool.poetry]
name = "Duelyst II Backend"
version = "0.1.0"
description = "Backend for everything Duelyst II related"
authors = ["Ghosterdriver"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
XlsxWriter = "^3.0.8"
"discord.py" = "^2.1.0"

[tool.poetry.dev-dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
